+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.2.0
+* No need to inject icons anymore
+
+1.1.0
+* Add support of rect,circle,line,polygon,ellipse tag of svg file
+
 1.0.8
 * Change the svg folder to src/svg